split up the SingleProduct component into many small components and imported all to the route through SingleProduct
