+++ src/components/BreadCrumbs/BreadCrumbs.module.css
@@ -1,0 +1,20 @@
+.breadcrumbs {
+  padding: 1rem;
+  font-size: 0.875rem;
+  color: var(--color-dark-gray);
+  display: flex;
+  gap: 0.5rem;
+}
+
+.crumb {
+  color: var(--color-dark-gray);
+  text-decoration: none;
+
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
+.active {
+  text-decoration: underline;
+}
