+++ src/components/ProductCards/SingleProduct/ProductDescription/ProductDescription.module.css
@@ -1,0 +1,61 @@
+/* .extendedInfoContainer {
+  padding: 2rem 1rem;
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+} */
+
+.infoHeader {
+  border-bottom: solid 1px var(--color-mid-gray);
+  margin-bottom: 1rem;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  cursor: pointer;
+  h2 {
+    color: var(--color-dark-gray);
+    text-transform: uppercase;
+    font-size: 1rem;
+    font-weight: 400;
+  }
+}
+
+.arrow {
+  border: solid var(--color-dark-gray);
+  border-width: 0 2px 2px 0;
+  display: inline-block;
+  width: 10px;
+  height: 10px;
+  transition: all 0.3s ease-in-out;
+}
+
+.arrowContainer {
+  padding: 0.2rem;
+}
+
+.up {
+  transform: rotate(-135deg);
+  -webkit-transform: rotate(-135deg);
+}
+
+.down {
+  transform: rotate(45deg);
+  -webkit-transform: rotate(45deg);
+}
+
+.hidden {
+  height: 0rem;
+  visibility: hidden;
+  opacity: 0;
+  transition: visibility 0s, opacity 0.5s ease-in-out;
+}
+
+.content {
+  transition: visibility 0s, opacity 0.5s ease-in-out;
+}
+
+.open {
+  height: 100%;
+  visibility: visible;
+  opacity: 1;
+}
