+++ src/components/ProductCards/SingleProduct/SimilarProducts/SimilarProducts.module.css
@@ -1,0 +1,18 @@
+.infoHeader {
+  border-bottom: solid 1px var(--color-mid-gray);
+  margin-bottom: 1rem;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  cursor: pointer;
+  h2 {
+    color: var(--color-dark-gray);
+    text-transform: uppercase;
+    font-size: 1rem;
+    font-weight: 400;
+  }
+}
+
+.similarProductsHeader {
+  margin-bottom: 2rem;
+}
