+++ src/components/ProductCards/SingleProduct/SimilarProducts/index.jsx
@@ -1,0 +1,13 @@
+import styles from "./SimilarProducts.module.css";
+import ProductsToDisplay from "../../ProductLists/index.jsx";
+
+export default function SimilarProducts({ tag }) {
+  return (
+    <div>
+      <div className={`${styles.infoHeader} ${styles.similarProductsHeader}`}>
+        <h2>Similar products</h2>
+      </div>
+      <ProductsToDisplay tag={tag} />
+    </div>
+  );
+}
