+++ src/components/ProductCards/SingleProduct/SingleProduct.module.css
@@ -1,0 +1,74 @@
+.productContainer {
+  @media screen and (min-width: 768px) {
+    display: flex;
+    padding: 1rem;
+  }
+}
+
+
+.imageWrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  /* max-height: 200px; */
+
+  img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    max-height: 400px;
+  }
+  @media screen and (min-width: 768px) {
+    height: 400px;
+  }
+}
+
+.productInfoContainer {
+  display: grid;
+  grid-template-rows: 1fr 1fr;
+  gap: 0.5rem;
+  width: 100%;
+  padding: 2rem 1rem;
+  background-color: var(--color-white);
+
+  h1 {
+    font-size: 1.125rem;
+    font-weight: 600;
+    color: var(--color-black);
+    @media screen and (min-width: 768px) {
+      font-size: 1.5rem;
+    }
+  }
+  @media screen and (min-width: 768px) {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    padding: 2rem;
+  }
+}
+
+.extendedInfoContainer {
+  padding: 2rem 1rem;
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+}
+
+
+.likeShareWrapper {
+  gap: 1rem;
+}
+
+.infoWrapper,
+.likeShareWrapper {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.infoWrapper1 {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 0.5rem;
+}
